refactor: delete CxError enum and use plural for factories provider name
