+++ core-libs/setup/ssr/error-handling/server-error-response-factory/index.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+export * from './cms-page-not-found-server-error-response-factory';
+export * from './provide-server-error-response-factories';
 export * from './server-error-response-factory';
 export * from './unknown-server-error-response-factory';
-export * from './cms-page-not-found-server-error-response-factory';
-export * from './provide-server-error-response-factory';
