+++ core-libs/setup/ssr/error-handling/server-error-response-factory/provide-server-error-response-factories.ts
@@ -6,13 +6,13 @@
 
 import { Provider } from '@angular/core';
 import { CmsPageNotFoundServerErrorResponseFactory } from './cms-page-not-found-server-error-response-factory';
-import { UnknownServerErrorResponseFactory } from './unknown-server-error-response-factory';
 import { SERVER_ERROR_RESPONSE_FACTORY } from './server-error-response-factory';
+import { UnknownServerErrorResponseFactory } from './unknown-server-error-response-factory';
 
 /**
  * This file provides the default server error response factories.
  */
-export const provideServerErrorResponseFactory = (): Provider[] => [
+export const provideServerErrorResponseFactories = (): Provider[] => [
   {
     provide: SERVER_ERROR_RESPONSE_FACTORY,
     useExisting: CmsPageNotFoundServerErrorResponseFactory,
