+++ core-libs/setup/ssr/error-handling/server-error-response/index.ts
@@ -4,5 +4,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './cx-server-error';
-export * from './cx-errors';
+export * from './cx-server-error-response';
