+++ core-libs/setup/ssr/providers/ssr-providers.ts
@@ -13,6 +13,7 @@ import {
 } from '@spartacus/core';
 
 import { ServerRespondingErrorHandler } from '../error-handling/multi-error-handlers';
+import { provideServerErrorResponseFactories } from '../error-handling/server-error-response-factory/provide-server-error-response-factories';
 import { getRequestOrigin } from '../express-utils/express-request-origin';
 import { getRequestUrl } from '../express-utils/express-request-url';
 import { serverLoggerServiceFactory } from '../logger';
@@ -20,7 +21,6 @@ import { REQUEST } from '../tokens/express.tokens';
 import { ServerOptions } from './model';
 import { serverRequestOriginFactory } from './server-request-origin';
 import { serverRequestUrlFactory } from './server-request-url';
-import { provideServerErrorResponseFactory } from '../error-handling/server-error-response-factory/provide-server-error-response-factory';
 
 /**
  * Returns the providers used for SSR and pre-rendering processes.
@@ -44,7 +44,7 @@ export function provideServer(options?: ServerOptions): Provider[] {
       useExisting: ServerRespondingErrorHandler,
       multi: true,
     },
-    provideServerErrorResponseFactory(),
+    provideServerErrorResponseFactories(),
   ];
 }
 /**
